Remove workarounds for upstream mlx-ruby/mlx-onnx fixes
Update mlx-ruby submodule to 476f721 (main) which includes:
- mx.array() dtype: keyword argument support (Issue #1)
- mx.mean() keepdims: parameter (Issue #2)
- Array#coerce for numeric-left ops (Issues #7, #8)
- update_modules_impl Module→Hash recursion fix (Issue #12)
- MLX_BUILD_SAFETENSORS=ON by default (Issue #5)

Update mlx-onnx submodule to 128d7de which adds GreaterEqual
ONNX lowering (Issue #14a). All 11 dense models now export to
ONNX with 100% node coverage.

Workarounds removed across lib/ and test/:
- .array(X).astype(T) → .array(X, dtype: T) in 50+ locations
- expand_dims(mean(x, axis), -1) → mean(x, axis, keepdims: true)
- Local update_modules_impl patch (now in upstream)

PRD updated to mark 8 issues as resolved, with remaining open
issues documented.

+++ lib/mlx_lm/benchmark.rb
@@ -6,9 +6,8 @@ module MlxLm
     def measure_generation(model, prompt_tokens: 32, gen_tokens: 64, vocab_size: 32000)
       mx = MLX::Core
 
-      # Create random prompt tokens (generate float then cast to int)
-      prompt = mx.random_uniform([prompt_tokens], 0.0, (vocab_size - 1).to_f, mx.float32)
-      prompt = prompt.astype(mx.int32)
+      # Create random prompt tokens
+      prompt = mx.random_uniform([prompt_tokens], 0.0, (vocab_size - 1).to_f, mx.float32).astype(mx.int32)
       mx.eval(prompt)
 
       # Create cache
